refactor: add LoginForm and SignUpForm components
